requirements.txt
aiogram==3.5.1
SQLAlchemy==2.0.20
asyncpg==0.29.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
